trying to fix linking issues

+++ test-test-server/include/physics/PhysicsSystem.h
@@ -307,7 +307,7 @@ public:
 
 	//delete all objects that are marked for deletion
     void deleteMarkedDynamicObjects() {
-		for (int i = movingObjects.size() - 1; i >= 0; i--) {
+		for (size_t i = movingObjects.size() - 1; i >= 0; i--) {
 			GameObject* obj = movingObjects[i];
 			if (obj->markDeleted) {
 				deleteDynamicObject(obj);
